refactor: Move policyTemplate.ts to fileTransformers folder
- Removed namespace from policyTemplate.ts
- Fixed imports
- @TODO: add documentation and move functionality around if needed
- Extracted decorators from policyTemplate.ts to decorators.d.ts
